% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyEnrichment_utils.R
\name{dotPlotEnrichment}
\alias{dotPlotEnrichment}
\title{Dot Plot Enrichment Visualization}
\usage{
dotPlotEnrichment(
  enrich,
  numSets = 10,
  colourBy = "p.adjust",
  titleForPlot = "Title",
  x_axis = "GeneRatio",
  textSize = 10,
  colour1 = "red",
  colour2 = "blue"
)
}
\arguments{
\item{enrich}{Enrichment result object (data.frame-like)}

\item{numSets}{Number of top results to show (default: 10)}

\item{colourBy}{Column name for color gradient (default: "p.adjust")}

\item{titleForPlot}{Plot title}

\item{x_axis}{X-axis variable ("GeneRatio" or "Count")}

\item{textSize}{Y-axis text size}

\item{colour1}{Low end color}

\item{colour2}{High end color}
}
\value{
plotly object
}
\description{
Creates interactive dot plot for enrichment results
}
