% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyEnrichment_utils.R
\name{heatPlotEnrichment}
\alias{heatPlotEnrichment}
\title{Heatmap Enrichment Visualization}
\usage{
heatPlotEnrichment(
  enrich,
  numSets = 10,
  colour = "black",
  sizeText = 10,
  title = NULL
)
}
\arguments{
\item{enrich}{Enrichment result object (e.g., from clusterProfiler)}

\item{numSets}{Number of top results to display (default: 10)}

\item{colour}{Tile color for gene presence (default: "black")}

\item{sizeText}{Axis text size (default: 10)}

\item{title}{Title of the plot}
}
\value{
Interactive plotly object
}
\description{
Creates interactive binary heatmap for enrichment analysis results showing gene presence/absence
}
\author{
Jacob Martin
}
