% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{shinyDESeq2}
\alias{shinyDESeq2}
\title{Backward compatibility wrapper for shinyDE (formerly shinyDESeq2)}
\usage{
shinyDESeq2(
  dds,
  res = NULL,
  coef = NULL,
  annot.by = NULL,
  use.lfcShrink = TRUE,
  lfcThreshold = 0,
  use.vst = TRUE,
  h.id = "ht1",
  genesets = NULL,
  swap.rownames = NULL,
  height = 800
)
}
\arguments{
\item{dds}{A \code{\link[DESeq2]{DESeqDataSet}} object.}

\item{res}{Either the object returned by \code{\link[DESeq2]{results}} or \code{\link[DESeq2]{lfcShrink}}
or a named list of such results.}

\item{coef}{A string indicating the coefficient name for which results will be generated.}

\item{annot.by}{A string or character vector containing the names of sample metadata variables.}

\item{use.lfcShrink}{Boolean indicating whether \code{\link[DESeq2]{lfcShrink}} should be used.}

\item{lfcThreshold}{Number passed to \code{\link[DESeq2]{lfcShrink}} or \code{\link[DESeq2]{results}}.}

\item{use.vst}{Boolean indicating whether \code{\link[DESeq2]{vst}} transformed counts should be used.}

\item{h.id}{String indicating unique ID for interactive heatmaps.}

\item{genesets}{Optional named list containing genesets.}

\item{swap.rownames}{String. The column name used to identify features.}

\item{height}{Number indicating height of app in pixels.}
}
\value{
A Shiny app.
}
\description{
This function provides backward compatibility for the old \code{shinyDESeq2} function.
It accepts a DESeqDataSet object and wraps it to work with the new generic \code{shinyDE} function.

For new code, consider using \code{\link{shinyDE}} directly, which accepts results from any DE analysis tool
(DESeq2, edgeR, limma, etc.).
}
\seealso{
\code{\link{shinyDE}}, \code{\link[DESeq2]{results}}, \code{\link[DESeq2]{lfcShrink}}, \code{\link[DESeq2]{resultsNames}}.
}
\author{
Jared Andrews, based on code by Zuguang Gu.
}
