% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAtools.R
\name{shinyPCAtools}
\alias{shinyPCAtools}
\title{Create an interactive Shiny app for PCAtools analysis and results exploration}
\usage{
shinyPCAtools(
  mat,
  metadata,
  removeVar = 0.3,
  scale = FALSE,
  center = TRUE,
  color.by = NULL,
  shape.by = NULL,
  annot.by = NULL,
  height = 850
)
}
\arguments{
\item{mat}{A matrix with features as rows and samples as columns.}

\item{metadata}{A dataframe containing sample metadata. The rownames must match the column names of the matrix.}

\item{removeVar}{Remove this \% of variables based on low variance.}

\item{scale}{Scale the data? Same as prcomp() 'scale' parameter.}

\item{center}{Center the data before performing PCA? Same as prcomp()
'center' parameter.}

\item{color.by}{A string containing the name of a sample metadata variable to be used to color points.}

\item{shape.by}{A string containing the name of a sample metadata variable to be used to shape points.}

\item{annot.by}{A string or character vector containing the names of sample metadata variables to be used for hover text.}

\item{height}{Number indicating height of app in pixels.}
}
\value{
A Shiny app wrapped around most PCAtools functions and plots.
}
\description{
Create an interactive Shiny app for PCAtools analysis and results exploration
}
\details{
Features with no variation will be removed prior to \code{\link[PCAtools]{pca}} being run.
}
\seealso{
\code{\link[PCAtools]{pca}}, \code{\link[PCAtools]{screeplot}},
\code{\link[PCAtools]{biplot}}, \code{\link[PCAtools]{pairsplot}}.
}
\author{
Jared Andrews
}
