% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyEnrichment_utils.R
\name{treeEnrichment}
\alias{treeEnrichment}
\title{Treemap Enrichment Visualization}
\usage{
treeEnrichment(reduceTerms, title = "Title", colourPalette = "Grey")
}
\arguments{
\item{reduceTerms}{Data frame of reduced enrichment terms with columns \code{parentTerm}, \code{term}, and \code{score}.}

\item{title}{Title of the treemap plot (default: "Title").}

\item{colourPalette}{a RColorBrewer::display.brewer.all() Colour pallete as a string}
}
\value{
A treemap object (list) returned by {treemap::treemap()}.
}
\description{
Generates a treemap of reduced enrichment terms, grouped by parent term and sized by score.
}
\author{
Jacob Martin
}
