% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyEnrichment_utils.R
\name{dotPlotEnrichment}
\alias{dotPlotEnrichment}
\title{DotPlot Function For Enrichment Results}
\usage{
dotPlotEnrichment(
  enrich,
  numSets = 10,
  colourBy = "p.adjust",
  titleForPlot = "Title",
  x_axis = "GeneRatio"
)
}
\arguments{
\item{enrich}{ClusterProfiler enrichment Object from enrichGO() function}

\item{numSets}{the number of gene sets you want to be displayed in the plot}

\item{colourBy}{Either p.adjust, qvalue or pvalue. Determines how the dots are coloured on a colour scale.}

\item{titleForPlot}{Title for the plot}

\item{x_axis}{Whether the x axis represents the GeneRatio or Count}
}
\description{
DotPlot Function For Enrichment Results
}
\author{
Jacob Martin
}
