% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECorr.R
\name{shinyDECorr}
\alias{shinyDECorr}
\title{Create an interactive Shiny app for correlation of differential expression results}
\usage{
shinyDECorr(
  res,
  sig.col = NULL,
  sig.thresh = 0.05,
  lfc.col = NULL,
  gene.col = NULL,
  expr.col = NULL,
  height = 800
)
}
\arguments{
\item{res}{A named list of data.frames containing differential expression analysis results.}

\item{sig.col}{String for the column name of the significance value, e.g. "padj". If not provided,
the function will search for commonly used values ("padj", "FDR", "svalue", "adj.P.Val") in the column names.}

\item{sig.thresh}{Number to be used as the significance threshold. Adjustable within the app.}

\item{lfc.col}{String for the column name of the log2 fold change column, e.g. "log2FC". If not provided,
the function will search for commonly used values ("log2FoldChange", "logFC") in the column names.}

\item{gene.col}{String for the column name containing the gene identifier. If not provided, rownames will
be used.}

\item{expr.col}{Optional string for the column name containing average expression. If not provided,
the function will search for commonly used values ("baseMean", "logCPM", "AveExpr") in the column names.}

\item{height}{Number indicating height of app in pixels.}
}
\value{
A Shiny app containing scatter plots comparing all combinations of the DE results to each other,
along with a line of best fit and correlation testing.
}
\description{
This shiny app generates scatter plots for every combination of DE results fed to it. It is useful for
comparing relative differences in differential expression for common conditions between different
backgrounds, groups, or settings.
}
\details{
Gene labels can be added to the plots by clicking a point. The labels can also be dragged around,
though adding labels to a plot will reset the positions, so it's recommended to add all labels prior to re-positioning them.

Comparisons will be limited to shared genes.
}
\author{
Jared Andrews
}
