% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R
\name{shinyDESeq2}
\alias{shinyDESeq2}
\title{Create an interactive Shiny app for DESeq2 analysis and results exploration}
\usage{
shinyDESeq2(
  dds,
  res = NULL,
  coef = NULL,
  annot.by = NULL,
  use.lfcShrink = TRUE,
  lfcThreshold = 0,
  use.vst = TRUE,
  samples.use = NULL,
  h.id = "ht1",
  height = 800
)
}
\arguments{
\item{dds}{A \code{\link[DESeq2]{DESeqDataSet}} object.}

\item{res}{The object returned by \code{\link[DESeq2]{results}} or \code{\link[DESeq2]{lfcShrink}} (recommended).
If not provided, it will be generated from the \code{dds} object via \code{\link[DESeq2]{lfcShrink}} and \code{coef}.}

\item{coef}{A string indicating the coefficient name for which results will be generated.
If not provided and \code{res} is \code{NULL}, the first non-intercept coefficient
provided by \code{\link[DESeq2]{resultsNames}} will be used.}

\item{annot.by}{A string or character vector containing the names of sample metadata variables to be used as heatmap annotations.
If not provided, all variables used in the \code{design} will be used.}

\item{use.lfcShrink}{Boolean indicating whether \code{\link[DESeq2]{lfcShrink}} should be used during results
generation. These are useful for ranking and visualization without the need for arbitrary filtering of low count genes.}

\item{lfcThreshold}{Number passed to \code{\link[DESeq2]{lfcShrink}} or \code{\link[DESeq2]{results}} to set an LFC threshold.
p-values, s-values, and adjusted p-values returned will be for whether the LFC is significantly greater in absolute value than the threshold.
Ignored if \code{res} is not \code{NULL}.}

\item{use.vst}{Boolean indicating whether \code{\link[DESeq2]{vst}} transformed counts should be used for heatmap generation (recommended). If \code{FALSE},
normalized counts will be used. The variance stabilizing transformation helps to reduce increased variance seen for low counts in log space.
This generally results in a better looking heatmap with fewer outliers.}

\item{samples.use}{Vector of Boolean values indicating which samples should be included in the visualizations.
Must be the same length as the number of samples in the \code{dds} object.}

\item{h.id}{String indicating unique ID for interactive heatmaps.
Required if multiple apps are run within the same Rmd file.}

\item{height}{Number indicating height of app in pixels.}
}
\value{
A Shiny app containing interconnected InteractiveComplexHeatmap, MAplot, and volcano plots along with full DE results.
}
\description{
This shiny app is composed of three tabs - an interactive heatmap, MA and volcano plots, and a table of full
differential expression results. The interactive heatmap will generate a sub-heatmap for selected rows/columns.
}
\details{
Gene labels can be added to the MAplot and volcano plot by clicking a point. The labels can also be dragged around,
though adding labels will reset the positions, so it's recommended to add all labels prior to re-positioning them.

Note that significance values of 0 will always be pushed above the y-limit of the volcano plot, as they are infinite values after log transformation.
}
\seealso{
\code{\link[DESeq2]{results}}, \code{\link[DESeq2]{lfcShrink}}, \code{\link[DESeq2]{resultsNames}}.
}
\author{
Jared Andrews, based on code by Zuguang Gu.
}
